optimize Crafter recipe merge

+++ src/main/java/tj/util/Counter.java
@@ -3,7 +3,7 @@ package tj.util;
 public final class Counter {
     private long value;
 
-    public Counter(int initialValue) {
+    public Counter(long initialValue) {
         this.value = initialValue;
     }
